requirements.txt
pymupdf4llm
PyPDF2
langchain
langchain-google-genai
google-generativeai
langchain-community
faiss-cpu
python-dotenv
streamlit
